refactor: RegService, RegController -> AuthService, AuthController 통합
