+++ src/main/java/com/davcatch/devcatch/controller/auth/RegRequest.java
@@ -1,4 +1,4 @@
-package com.davcatch.devcatch.controller.registration;
+package com.davcatch.devcatch.controller.auth;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
@@ -10,7 +10,10 @@ public class RegRequest {
 	@NotBlank(message = "이름은 필수 입력값 입니다.")
 	private String name;
 
-	@NotBlank
+	@NotBlank(message = "이메일은 필수 입력값 입니다")
 	@Email
 	private String email;
+
+	@NotBlank(message = "비밀번호는 필수 입력값입니다.")
+	private String password;
 }
